requirements.txt
selenium==4.35.0
allure-pytest==2.15.0
pytest==8.4.2
requests==2.32.5
webdriver-manager==4.0.2